pyproject.toml
[project]
name = "opensearch-mcp-server-py"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiohttp>=3.11.18",
    "boto3>=1.38.3",
    "mcp[cli]>=1.6.0",
    "opensearch-py>=2.8.0",
    "pydantic>=2.11.3",
    "pyyaml>=6.0.2",
    "requests-aws4auth>=1.3.1",
    "semver>=3.0.4",
]

license = "Apache-2.0"
license-files = ["LICENSE", "NOTICE" ]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "pytest-cov>=6.2.1",
    "ruff>=0.9.7",
]

[build-system]
requires = ["setuptools >= 77.0.3"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 99
extend-include = ["*.ipynb"]
exclude = [
    ".venv",
    "**/__pycache__",
    "**/node_modules",
    "**/dist",
    "**/build",
    "**/env",
    "**/.ruff_cache",
    "**/.venv",
    "**/.ipynb_checkpoints"
]
force-exclude = true

[tool.ruff.lint]
exclude = ["__init__.py"]
select = ["C", "D", "E", "F", "I", "W"]
ignore = ["C901", "E501", "E741", "F402", "F823", "D100", "D106"]

[tool.ruff.lint.isort]
lines-after-imports = 2
no-sections = true

[tool.ruff.lint.per-file-ignores]
"**/*.ipynb" = ["F704"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.format]
quote-style = "single"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = true

[project.scripts]
opensearch-mcp-server-py = "mcp_server_opensearch:main"  # Importable path

[project.urls]
Homepage = "https://github.com/opensearch-project/opensearch-mcp-server-py"
Documentation = "https://github.com/opensearch-project/opensearch-mcp-server-py/blob/main/USER_GUIDE.md"
Source = "https://github.com/opensearch-project/opensearch-mcp-server-py.git"
"Bug Tracker" = "https://github.com/opensearch-project/opensearch-mcp-server-py/issues"
Changelog = "https://github.com/opensearch-project/opensearch-mcp-server-py/blob/main/CHANGELOG.md"
